feature/ON4T-1feature/ON4T-174_Backend_for_Tasks_DeleteTasks - general outline for the delete mapping in the task REST controller - deletion via objectID provided by mongoDB

+++ src/main/java/com/gocaspi/taskfly/Task/TaskController.java
@@ -81,9 +81,21 @@ public class TaskController {
         Task[] response = RemoveNullElements(tasksOfUser);
         return new Gson().toJson(response);
     }
+
+    /**
+     *
+     * @param id
+     */
     @DeleteMapping("/{id}")
-    public void deletePerson(@PathVariable String id){
-        repository.deleteById(id);
+    public String deleteTask(@PathVariable String id){
+        if(repository.existsById(id)){
+            repository.deleteById(id);
+            return new String("Deleted "+id+" successfully");
+        }
+        else {
+            return new String("could not find matching Task to the provided id");
+        }
+
     }
 
     /**
